added; support for continueOnError
Web apps often still need to handle input errors in
their routes. This commit adds a `continueOnError` validation
option (defaulting to false) which when enabled, causes validation
middleware to continue iteration, passing through control to
the defined route.

In the route, we can determine if input validation failed by
checking for the existence of `ctx.invalid`. If it exists,
validation failed and it is set to an object containing the
details of the failure.

    ctx.invalid = {
      headers: ..
      params: ..
      query: ..
      body: ..
      type: ..
    }